Rename *TieredCommits* test classes
Fixes #2031
